Fix offline mode website

+++ src/store/profile/action.js
@@ -2,6 +2,7 @@ import { ENDPOINTS } from "../../common/endpoints";
 import useSnackbar from "../../shared/hooks/useSnackbar";
 import { ACTION } from "./reducer";
 import uuid from "react-uuid";
+import { loadState } from "../localStorage";
 
 const setProfileId = (profileId) => (dispatch) => {
   dispatch({
@@ -10,54 +11,65 @@ const setProfileId = (profileId) => (dispatch) => {
   });
 };
 
-export const getProfileById = (profileId) => async (dispatch) => {
+export const getProfileById = (profileId) => async (dispatch, getState) => {
   dispatch(setIsLoading(true));
-  // const [error] = useSnackbar("error");
-  if (profileId === undefined) {
-    await fetch(ENDPOINTS.URL)
-      .then((resp) => {
-        if (resp.status >= 400) {
-          console.log(resp);
-        } else return resp.json();
-      })
-      .then((resp) => {
-        dispatch({
-          type: ACTION.GET_PROFILE_BY_PID,
-          userProfile: resp.profile,
-        });
-        dispatch(setProfileId(resp.profile.id));
-      })
-      .catch((err) => {
-        console.log(err);
-        dispatch({
-          type: ACTION.GET_PROFILE_BY_PID,
-          userProfile: undefined,
-        });
-      });
+  const { isOffline } = getState().profile;
+
+  if (isOffline) {
+    const storedData = loadState();
+    console.log(storedData.profile);
+    dispatch({
+      type: ACTION.GET_PROFILE_BY_PID,
+      userProfile: storedData.profile.userProfile,
+    });
   } else {
-    await fetch(`${ENDPOINTS.URL}/${profileId}`)
-      .then((resp) => {
-        if (resp.status >= 400) {
+    if (profileId === undefined) {
+      await fetch(ENDPOINTS.URL)
+        .then((resp) => {
+          if (resp.status >= 400) {
+            console.log(resp);
+          } else return resp.json();
+        })
+        .then((resp) => {
+          dispatch({
+            type: ACTION.GET_PROFILE_BY_PID,
+            userProfile: resp.profile,
+          });
+          dispatch(setProfileId(resp.profile.id));
+        })
+        .catch((err) => {
+          console.log(err);
+          console.log("userProfile is set to undefined!");
+          dispatch({
+            type: ACTION.GET_PROFILE_BY_PID,
+            userProfile: undefined,
+          });
+        });
+    } else {
+      await fetch(`${ENDPOINTS.URL}/${profileId}`)
+        .then((resp) => {
+          if (resp.status >= 400) {
+            console.log(resp);
+          } else {
+            return resp.json();
+          }
+        })
+        .then((resp) => {
           console.log(resp);
-          // error("No profiles found, create one!");
-        } else {
-          return resp.json();
-        }
-      })
-      .then((resp) => {
-        console.log(resp);
-        dispatch({
-          type: ACTION.GET_PROFILE_BY_PID,
-          userProfile: resp.profile,
+          dispatch({
+            type: ACTION.GET_PROFILE_BY_PID,
+            userProfile: resp.profile,
+          });
+        })
+        .catch((err) => {
+          console.log(err);
+          console.log("Set userProfile to undefined!");
+          dispatch({
+            type: ACTION.GET_PROFILE_BY_PID,
+            userProfile: undefined,
+          });
         });
-      })
-      .catch((err) => {
-        console.log(err);
-        dispatch({
-          type: ACTION.GET_PROFILE_BY_PID,
-          userProfile: undefined,
-        });
-      });
+    }
   }
   dispatch(setIsLoading(false));
 };
@@ -85,29 +97,42 @@ export const getWorkExperienceById = (profileId, weId) => async (dispatch) => {
   dispatch(setIsLoading(false));
 };
 
-export const getAllWorkExperiences = (profileId) => async (dispatch) => {
+export const getAllWorkExperiences = (profileId) => async (
+  dispatch,
+  getState
+) => {
   dispatch(setIsLoading(true));
   const [error] = useSnackbar("error");
+  const { isOffline } = getState().profile;
 
-  // /:profileId/workExperience/all
-  await fetch(`${ENDPOINTS.URL}/${profileId}/workExperience/all`)
-    .then((resp) => {
-      if (resp.status >= 400) {
-        console.log(resp);
-        throw error(
-          "Failed to get your work experiences, please try again later."
-        );
-      } else {
-        return resp.json();
-      }
-    })
-    .then((resp) => {
-      dispatch({
-        type: ACTION.GET_ALL_WORK_EXPERIENCE_BY_PID,
-        allWorkExperiences: resp.workExperiences,
-      });
-    })
-    .catch((err) => console.log(err));
+  if (isOffline) {
+    const storedData = loadState();
+    console.log(storedData.profile);
+    dispatch({
+      type: ACTION.GET_ALL_WORK_EXPERIENCE_BY_PID,
+      allWorkExperiences: storedData.profile.userProfile.workExperiences,
+    });
+  } else {
+    // /:profileId/workExperience/all
+    await fetch(`${ENDPOINTS.URL}/${profileId}/workExperience/all`)
+      .then((resp) => {
+        if (resp.status >= 400) {
+          console.log(resp);
+          throw error(
+            "Failed to get your work experiences, please try again later."
+          );
+        } else {
+          return resp.json();
+        }
+      })
+      .then((resp) => {
+        dispatch({
+          type: ACTION.GET_ALL_WORK_EXPERIENCE_BY_PID,
+          allWorkExperiences: resp.workExperiences,
+        });
+      })
+      .catch((err) => console.log(err));
+  }
   dispatch(setIsLoading(false));
 };
 
@@ -244,32 +269,40 @@ export const createWorkExperience = (profileId, workExperience) => async (
     .catch((err) => console.log(err));
   dispatch(setIsLoading(false));
 };
-export const deleteWorkExperience = (profileId, weId) => async (dispatch) => {
+
+export const deleteWorkExperience = (profileId, weId) => async (
+  dispatch,
+  getState
+) => {
   dispatch(setIsLoading(true));
   const [error] = useSnackbar("error");
   const [success] = useSnackbar("success");
+  const { isOffline } = getState().profile;
 
-  ///:profileId/workExperience/:workExperienceId
-  await fetch(`${ENDPOINTS.URL}/${profileId}/workExperience/${weId}`, {
-    method: "DELETE",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  })
-    .then((resp) => {
-      if (resp.status >= 400) {
+  if (isOffline) {
+  } else {
+    ///:profileId/workExperience/:workExperienceId
+    await fetch(`${ENDPOINTS.URL}/${profileId}/workExperience/${weId}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+      .then((resp) => {
+        if (resp.status >= 400) {
+          console.log(resp);
+          throw error("Failed to delete work experience, please try again.");
+        } else {
+          return resp.json();
+        }
+      })
+      .then((resp) => {
         console.log(resp);
-        throw error("Failed to delete work experience, please try again.");
-      } else {
-        return resp.json();
-      }
-    })
-    .then((resp) => {
-      console.log(resp);
-      success("Successfully deleted work experience!");
-      dispatch(getAllWorkExperiences(profileId));
-    })
-    .catch((err) => console.log(err));
+        success("Successfully deleted work experience!");
+        dispatch(getAllWorkExperiences(profileId));
+      })
+      .catch((err) => console.log(err));
+  }
   dispatch(setIsLoading(false));
 };
 
@@ -288,4 +321,49 @@ export const resetSelectedWorkExperience = () => (dispatch) => {
 
 export const setIsLoading = (desiredState) => (dispatch) => {
   dispatch({ type: ACTION.SET_IS_LOADING, isLoading: desiredState });
+};
+
+export const setIsOffline = (isOffline) => (dispatch) => {
+  console.log(isOffline);
+  dispatch({ type: ACTION.SET_IS_OFFLINE, isOffline: isOffline });
+};
+
+// request = {
+// endpoint: string
+// method: 'PATCH', 'POST', 'DELETE'
+// body: {} || null
+// }
+export const setOfflineQueue = (request) => (dispatch, getState) => {
+  const { offlineQueue } = getState().profile;
+
+  dispatch({
+    type: ACTION.SET_OFFLINE_QUEUE,
+    offlineQueue: offlineQueue.push(request),
+  });
+};
+
+export const sendRequests = (allOfflineRequests) => {
+  allOfflineRequests.map(async (req) => {
+    if (req.method === "DELETE") {
+      await fetch(req.endpoint, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+    } else if (req.method === ("PATCH" || "POST")) {
+      await fetch(req.endpoint, {
+        method: req.method,
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: req.body,
+      });
+    } else {
+      console.log(req); //ignore request
+      return;
+    }
+  });
+  //update stores
+  // dispatch(getProfileById())
 };
